requirements.txt
bitarray==2.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.5
Deprecated==1.2.14
fonttools==4.44.0
idna==3.6
kiwisolver==1.4.5
lark==1.1.8
lxml==5.1.0
marko==2.0.3
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pybloom-live==4.0.0
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
seaborn==0.13.0
six==1.16.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
wrapt==1.16.0
xxhash==3.4.1
